+++ CHANGELOG.md
@@ -8,3 +8,11 @@ All notable changes to this project will be documented in this file.
 - Automatic local HTTPS to server SSH URL compilation logic.
 - Self-contained dynamic GitHub Deploy Key token injector.
 - Circuit-breaker Laravel production performance optimization queue wrapper over SSH.
+
+## [1.0.1] - 2026-06-08
+### Added
+- Detailed step by step guide Readme.md
+
+## [1.1.1] - 2026-06-09
+## Added
+- Fixed the storage:link error and debug log